Use 'conda env create' instead of 'conda list' to get dependencies
