+++ test/Conda/testdata/conda-env-create-out.json
@@ -7,40 +7,7 @@
   "dependencies": [
     "conda-forge/osx-64::bzip2==1.0.8=h0d85af4_4",
     "defaults/osx-64::ca-certificates==2022.07.19=hecd8cb5_0",
-    "conda-forge/osx-64::libcxx==14.0.6=hccf4f1f_0",
-    "conda-forge/osx-64::libffi==3.4.2=h0d85af4_5",
-    "conda-forge/osx-64::libzlib==1.2.12=hfd90126_3",
-    "conda-forge/osx-64::llvm-openmp==14.0.4=ha654fa7_0",
-    "conda-forge/osx-64::ncurses==6.3=h96cf925_1",
-    "conda-forge/noarch::tzdata==2022c=h191b570_0",
-    "conda-forge/osx-64::xz==5.2.6=h775f41a_0",
-    "conda-forge/osx-64::eigen==3.4.0=h940c156_0",
-    "conda-forge/osx-64::fmt==9.1.0=hb8565cd_0",
-    "conda-forge/osx-64::icu==70.1=h96cf925_0",
-    "conda-forge/osx-64::libgfortran5==11.3.0=h082f757_25",
-    "conda-forge/osx-64::libsqlite==3.39.3=ha978bb4_0",
-    "conda-forge/osx-64::openssl==3.0.5=hfd90126_2",
-    "conda-forge/osx-64::rapidjson==1.1.0=hb1e8313_1002",
-    "conda-forge/osx-64::readline==8.1.2=h3899abd_0",
-    "conda-forge/osx-64::spdlog==1.10.0=hbb4e6a2_0",
-    "conda-forge/osx-64::tk==8.6.12=h5dbffcc_0",
-    "conda-forge/osx-64::zlib==1.2.12=hfd90126_3",
-    "conda-forge/osx-64::zstd==1.5.2=hfa58983_4",
-    "conda-forge/osx-64::boost-cpp==1.78.0=h8b082ac_1",
-    "conda-forge/osx-64::libgfortran==5.0.0=10_4_0_h97931a8_25",
-    "conda-forge/osx-64::python==3.10.6=hc14f532_0_cpython",
-    "defaults/osx-64::absl-py==1.3.0=py310hecd8cb5_0",
-    "defaults/osx-64::aenum==3.1.5=py310hecd8cb5_0",
-    "conda-forge/osx-64::libopenblas==0.3.21=openmp_h429af6e_3",
-    "conda-forge/osx-64::python_abi==3.10=2_cp310",
-    "conda-forge/noarch::setuptools==65.3.0=pyhd8ed1ab_1",
-    "conda-forge/noarch::wheel==0.37.1=pyhd8ed1ab_0",
-    "conda-forge/osx-64::libblas==3.9.0=16_osx64_openblas",
-    "conda-forge/noarch::pip==22.2.2=pyhd8ed1ab_0",
-    "conda-forge/osx-64::libcblas==3.9.0=16_osx64_openblas",
-    "conda-forge/osx-64::liblapack==3.9.0=16_osx64_openblas",
-    "conda-forge/osx-64::numpy==1.23.3=py310h1b7c290_0",
-    "conda-forge/osx-64::boost==1.78.0=py310h84c6d00_0",
-    "conda-forge/osx-64::vowpalwabbit==9.3.0=py310h630de0f_2"
+    "pkgs/main/noarch::tzdata==2022c=h191b570_0",
+    "unparsable dep string"
   ]
 }
