--- test/Conda/testdata/conda-list-output.txt
@@ -1,32 +1,0 @@
-[
-  {
-    "base_url": "https://repo.anaconda.com/pkgs/main",
-    "build_number": 0,
-    "build_string": "py38haf1e3a3_0",
-    "channel": "pkgs/main",
-    "dist_name": "biopython-1.78-py38haf1e3a3_0",
-    "name": "biopython",
-    "platform": "osx-64",
-    "version": "1.78"
-  },
-  {
-    "base_url": "https://repo.anaconda.com/pkgs/main",
-    "build_number": 1,
-    "build_string": "mkl",
-    "channel": "pkgs/main",
-    "dist_name": "blas-1.0-mkl",
-    "name": "blas",
-    "platform": "osx-64",
-    "version": "1.0"
-  },
-  {
-    "base_url": "https://repo.anaconda.com/pkgs/main",
-    "build_number": 1,
-    "build_string": "hecd8cb5_1",
-    "channel": "pkgs/main",
-    "dist_name": "ca-certificates-2021.1.19-hecd8cb5_1",
-    "name": "ca-certificates",
-    "platform": "osx-64",
-    "version": "2021.1.19"
-  }
-]
